Couple of fixes to make things work on Mac OS: - use the default graphic renderer. Quartz is no longer defined it seems. - avoid calling glCalls() when SDL claims the window is usable with gl. There   might not be a gl context.

+++ src/ui_fsmenu/options.cc
@@ -341,7 +341,8 @@ FullscreenMenuOptions::FullscreenMenuOptions
 		if (800 <= mode.w && 600 <= mode.h &&
 			 (SDL_BITSPERPIXEL(mode.format) == 32 ||
 			  SDL_BITSPERPIXEL(mode.format) == 24)) {
-			ScreenResolution this_res = {mode.w, mode.h, SDL_BITSPERPIXEL(mode.format)};
+			ScreenResolution this_res = {
+			   mode.w, mode.h, static_cast<int32_t>(SDL_BITSPERPIXEL(mode.format))};
 			if (this_res.depth == 24) this_res.depth = 32;
 			if (m_resolutions.empty()
 				 || this_res.xres != m_resolutions.rbegin()->xres
